@@ -1,25 +1,29 @@
 [package]
 name = "czkawka_cli"
-version = "6.1.0"
+version = "7.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.70.0"
+rust-version = "1.74.0"
 description = "CLI frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
 repository = "https://github.com/qarmin/czkawka"
 
 [dependencies]
-clap = { version = "4.3", features = ["derive"] }
+clap = { version = "4.5", features = ["derive"] }
 
 # For enum types
 image_hasher = "1.2"
 
 log = "0.4.20"
 handsome_logger = "0.8"
-fun_time = { version = "0.3.1", features = ["log"] }
-czkawka_core = { path = "../czkawka_core", version = "6.1.0", features = [] }
+fun_time = { version = "0.3", features = ["log"] }
+czkawka_core = { path = "../czkawka_core", version = "7.0.0", features = [] }
+indicatif = "0.17"
+crossbeam-channel = { version = "0.5", features = [] }
+ctrlc = { version = "3.4", features = ["termination"] }
 
 [features]
 default = []
 heif = ["czkawka_core/heif"]
+libraw = ["czkawka_core/libraw"]
