@@ -1,26 +1,24 @@
 [package]
 name = "czkawka_gui"
-version = "6.1.0"
+version = "7.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
 edition = "2021"
-rust-version = "1.70.0"
+rust-version = "1.74.0"
 description = "GTK frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
 repository = "https://github.com/qarmin/czkawka"
 
 [dependencies]
-gdk4 = "0.7"
-glib = "0.18"
+gdk4 = "0.8"
+glib = "0.19"
+gtk4 = { version = "0.8", default-features = false, features = ["v4_6"] }
 
 humansize = "2.1"
-chrono = "0.4.31"
+chrono = "0.4.34"
 
 # Used for sending stop signal across threads
-crossbeam-channel = "0.5.8"
-
-# To get information about progress
-futures = "0.3.28"
+crossbeam-channel = "0.5"
 
 # For saving/loading config files to specific directories
 directories-next = "2.0"
@@ -32,29 +30,29 @@ open = "5.0"
 image = "0.24"
 
 # To be able to use custom select
-regex = "1.9"
+regex = "1.10"
 
 # To get image_hasher types
 image_hasher = "1.2"
 
 # Move files to trash
-trash = "3.0"
+trash = "3.3"
 
 # For moving files(why std::fs doesn't have such features?)
 fs_extra = "1.3"
 
 # Language
 i18n-embed = { version = "0.14", features = ["fluent-system", "desktop-requester"] }
-i18n-embed-fl = "0.7"
-rust-embed = "8.0"
-once_cell = "1.18"
+i18n-embed-fl = "0.8"
+rust-embed = { version = "8.2", features = ["debug-embed"] }
+once_cell = "1.19"
 
 log = "0.4.20"
 handsome_logger = "0.8"
-fun_time = { version = "0.3.1", features = ["log"] }
+fun_time = { version = "0.3", features = ["log"] }
+rayon = "1.8"
 
-czkawka_core = { path = "../czkawka_core", version = "6.1.0", features = [] }
-gtk4 = { version = "0.7", default-features = false, features = ["v4_6"] }
+czkawka_core = { path = "../czkawka_core", version = "7.0.0", features = [] }
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.9", features = ["combaseapi", "objbase", "shobjidl_core", "windef", "winerror", "wtypesbase", "winuser"] }
@@ -62,3 +60,4 @@ winapi = { version = "0.3.9", features = ["combaseapi", "objbase", "shobjidl_cor
 [features]
 default = []
 heif = ["czkawka_core/heif"]
+libraw = ["czkawka_core/libraw"]
