@@ -1,15 +1,12 @@
-use std::collections::BTreeMap;
-use std::fs::{DirEntry, File, Metadata};
+use std::collections::{BTreeMap, HashSet};
+use std::fs::File;
 use std::io::prelude::*;
-
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::sync::Arc;
+use std::sync::atomic::Ordering;
 use std::{fs, mem, panic};
 
-use crossbeam_channel::Receiver;
+use crossbeam_channel::{Receiver, Sender};
 use fun_time::fun_time;
-use futures::channel::mpsc::UnboundedSender;
 use log::debug;
 use pdf::file::FileOptions;
 use pdf::object::ParseOptions;
@@ -19,23 +16,23 @@ use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::common::{
-    check_folder_children, check_if_stop_received, create_crash_message, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads, AUDIO_FILES_EXTENSIONS,
+    check_if_stop_received, create_crash_message, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads, AUDIO_FILES_EXTENSIONS,
     IMAGE_RS_BROKEN_FILES_EXTENSIONS, PDF_FILES_EXTENSIONS, ZIP_FILES_EXTENSIONS,
 };
 use crate::common_cache::{get_broken_files_cache_file, load_cache_from_file_generalized_by_path, save_cache_to_file_generalized};
-use crate::common_dir_traversal::{common_get_entry_data_metadata, common_read_dir, get_lowercase_name, get_modified_time, CheckingMethod, ProgressData, ToolType};
+use crate::common_dir_traversal::{CheckingMethod, DirTraversalBuilder, DirTraversalResult, FileEntry, ProgressData, ToolType};
 use crate::common_tool::{CommonData, CommonToolData, DeleteMethod};
 use crate::common_traits::*;
 
-#[derive(Clone, Serialize, Deserialize)]
-pub struct FileEntry {
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct BrokenEntry {
     pub path: PathBuf,
     pub modified_date: u64,
     pub size: u64,
     pub type_of_file: TypeOfFile,
     pub error_string: String,
 }
-impl ResultEntry for FileEntry {
+impl ResultEntry for BrokenEntry {
     fn get_path(&self) -> &Path {
         &self.path
     }
@@ -47,6 +44,19 @@ impl ResultEntry for FileEntry {
     }
 }
 
+impl FileEntry {
+    fn into_broken_entry(self) -> BrokenEntry {
+        BrokenEntry {
+            size: self.size,
+            path: self.path,
+            modified_date: self.modified_date,
+
+            type_of_file: TypeOfFile::Unknown,
+            error_string: String::new(),
+        }
+    }
+}
+
 #[derive(Copy, Clone, PartialEq, Eq, Serialize, Deserialize, Debug)]
 pub enum TypeOfFile {
     Unknown = -1,
@@ -55,6 +65,8 @@ pub enum TypeOfFile {
     Audio,
     PDF,
 }
+
+const MAX_BROKEN_FILES_STAGE: u8 = 1;
 
 bitflags! {
     #[derive(PartialEq, Copy, Clone)]
@@ -76,8 +88,8 @@ pub struct Info {
 pub struct BrokenFiles {
     common_data: CommonToolData,
     information: Info,
-    files_to_check: BTreeMap<String, FileEntry>,
-    broken_files: Vec<FileEntry>,
+    files_to_check: BTreeMap<String, BrokenEntry>,
+    broken_files: Vec<BrokenEntry>,
     checked_types: CheckedTypes,
 }
 
@@ -93,8 +105,8 @@ impl BrokenFiles {
     }
 
     #[fun_time(message = "find_broken_files", level = "info")]
-    pub fn find_broken_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) {
-        self.optimize_dirs_before_start();
+    pub fn find_broken_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) {
+        self.prepare_items();
         if !self.check_files(stop_receiver, progress_sender) {
             self.common_data.stopped_search = true;
             return;
@@ -108,122 +120,63 @@ impl BrokenFiles {
     }
 
     #[fun_time(message = "check_files", level = "debug")]
-    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
-        let mut folders_to_check: Vec<PathBuf> = Vec::with_capacity(1024 * 2); // This should be small enough too not see to big difference and big enough to store most of paths without needing to resize vector
+    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
+        let zip_extensions = ZIP_FILES_EXTENSIONS.iter().collect::<HashSet<_>>();
+        let audio_extensions = AUDIO_FILES_EXTENSIONS.iter().collect::<HashSet<_>>();
+        let pdf_extensions = PDF_FILES_EXTENSIONS.iter().collect::<HashSet<_>>();
+        let images_extensions = IMAGE_RS_BROKEN_FILES_EXTENSIONS.iter().collect::<HashSet<_>>();
 
-        // Add root folders for finding
-        for id in &self.common_data.directories.included_directories {
-            folders_to_check.push(id.clone());
+        let mut extensions = Vec::new();
+        let vec_extensions = [
+            (CheckedTypes::PDF, PDF_FILES_EXTENSIONS),
+            (CheckedTypes::AUDIO, AUDIO_FILES_EXTENSIONS),
+            (CheckedTypes::ARCHIVE, ZIP_FILES_EXTENSIONS),
+            (CheckedTypes::IMAGE, IMAGE_RS_BROKEN_FILES_EXTENSIONS),
+        ];
+        for (checked_type, extensions_to_add) in &vec_extensions {
+            if self.checked_types.contains(*checked_type) {
+                extensions.extend_from_slice(extensions_to_add);
+            }
         }
 
-        let (progress_thread_handle, progress_thread_run, atomic_counter, _check_was_stopped) =
-            prepare_thread_handler_common(progress_sender, 0, 1, 0, CheckingMethod::None, self.common_data.tool_type);
+        self.common_data.extensions.set_and_validate_allowed_extensions(&extensions);
+        if !self.common_data.extensions.set_any_extensions() {
+            return true;
+        }
 
-        debug!("check_files - starting to collect files");
-        while !folders_to_check.is_empty() {
-            if check_if_stop_received(stop_receiver) {
-                send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
-                return false;
-            }
+        let result = DirTraversalBuilder::new()
+            .group_by(|_fe| ())
+            .stop_receiver(stop_receiver)
+            .progress_sender(progress_sender)
+            .common_data(&self.common_data)
+            .max_stage(MAX_BROKEN_FILES_STAGE)
+            .build()
+            .run();
 
-            let segments: Vec<_> = folders_to_check
-                .par_iter()
-                .map(|current_folder| {
-                    let mut dir_result = vec![];
-                    let mut warnings = vec![];
-                    let mut fe_result = vec![];
-
-                    let Some(read_dir) = common_read_dir(current_folder, &mut warnings) else {
-                        return (dir_result, warnings, fe_result);
-                    };
-
-                    // Check every sub folder/file/link etc.
-                    for entry in read_dir {
-                        let Some((entry_data, metadata)) = common_get_entry_data_metadata(&entry, &mut warnings, current_folder) else {
-                            continue;
-                        };
-
-                        if metadata.is_dir() {
-                            check_folder_children(
-                                &mut dir_result,
-                                &mut warnings,
-                                current_folder,
-                                entry_data,
-                                self.common_data.recursive_search,
-                                &self.common_data.directories,
-                                &self.common_data.excluded_items,
-                            );
-                        } else if metadata.is_file() {
-                            if let Some(file_entry) = self.get_file_entry(&metadata, &atomic_counter, entry_data, &mut warnings, current_folder) {
-                                fe_result.push((file_entry.path.to_string_lossy().to_string(), file_entry));
-                            }
-                        }
-                    }
-                    (dir_result, warnings, fe_result)
-                })
-                .collect();
-            debug!("check_files - collected files");
-
-            // Advance the frontier
-            folders_to_check.clear();
-
-            // Process collected data
-            for (segment, warnings, fe_result) in segments {
-                folders_to_check.extend(segment);
+        match result {
+            DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
+                self.files_to_check = grouped_file_entries
+                    .into_values()
+                    .flatten()
+                    .map(|fe| {
+                        let mut broken_entry = fe.into_broken_entry();
+                        broken_entry.type_of_file = check_extension_availability(broken_entry.get_path(), &images_extensions, &zip_extensions, &audio_extensions, &pdf_extensions);
+                        (broken_entry.path.to_string_lossy().to_string(), broken_entry)
+                    })
+                    .collect();
                 self.common_data.text_messages.warnings.extend(warnings);
-                for (name, fe) in fe_result {
-                    self.files_to_check.insert(name, fe);
-                }
+                debug!("check_files - Found {} files to check.", self.files_to_check.len());
+                true
             }
-        }
 
-        send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
-        true
+            DirTraversalResult::Stopped => false,
+        }
     }
 
-    fn get_file_entry(
-        &self,
-        metadata: &Metadata,
-        atomic_counter: &Arc<AtomicUsize>,
-        entry_data: &DirEntry,
-        warnings: &mut Vec<String>,
-        current_folder: &Path,
-    ) -> Option<FileEntry> {
-        atomic_counter.fetch_add(1, Ordering::Relaxed);
-
-        let Some(file_name_lowercase) = get_lowercase_name(entry_data, warnings) else {
-            return None;
-        };
-
-        if !self.common_data.allowed_extensions.matches_filename(&file_name_lowercase) {
-            return None;
-        }
-
-        let type_of_file = check_extension_availability(&file_name_lowercase);
-
-        if !check_if_file_extension_is_allowed(&type_of_file, &self.checked_types) {
-            return None;
-        }
-
-        let current_file_name = current_folder.join(entry_data.file_name());
-        if self.common_data.excluded_items.is_excluded(&current_file_name) {
-            return None;
-        }
-
-        let fe: FileEntry = FileEntry {
-            path: current_file_name.clone(),
-            modified_date: get_modified_time(metadata, warnings, &current_file_name, false),
-            size: metadata.len(),
-            type_of_file,
-            error_string: String::new(),
-        };
-        Some(fe)
-    }
-
-    fn check_broken_image(&self, mut file_entry: FileEntry) -> Option<FileEntry> {
+    fn check_broken_image(&self, mut file_entry: BrokenEntry) -> Option<BrokenEntry> {
         let mut file_entry_clone = file_entry.clone();
 
-        let result = panic::catch_unwind(|| {
+        panic::catch_unwind(|| {
             if let Err(e) = image::open(&file_entry.path) {
                 let error_string = e.to_string();
                 // This error is a problem with image library, remove check when https://github.com/image-rs/jpeg-decoder/issues/130 will be fixed
@@ -233,19 +186,15 @@ impl BrokenFiles {
                 file_entry.error_string = error_string;
             }
             Some(file_entry)
-        });
-
-        // If image crashed during opening, needs to be printed info about crashes thing
-        if let Ok(image_result) = result {
-            image_result
-        } else {
+        })
+        .unwrap_or_else(|_| {
             let message = create_crash_message("Image-rs", &file_entry_clone.path.to_string_lossy(), "https://github.com/Serial-ATA/lofty-rs");
             println!("{message}");
             file_entry_clone.error_string = message;
             Some(file_entry_clone)
-        }
+        })
     }
-    fn check_broken_zip(&self, mut file_entry: FileEntry) -> Option<FileEntry> {
+    fn check_broken_zip(&self, mut file_entry: BrokenEntry) -> Option<BrokenEntry> {
         match File::open(&file_entry.path) {
             Ok(file) => {
                 if let Err(e) = zip::ZipArchive::new(file) {
@@ -256,35 +205,32 @@ impl BrokenFiles {
             Err(_inspected) => None,
         }
     }
-    fn check_broken_audio(&self, mut file_entry: FileEntry) -> Option<FileEntry> {
+    fn check_broken_audio(&self, mut file_entry: BrokenEntry) -> Option<BrokenEntry> {
         match File::open(&file_entry.path) {
             Ok(file) => {
                 let mut file_entry_clone = file_entry.clone();
 
-                let result = panic::catch_unwind(|| {
+                panic::catch_unwind(|| {
                     if let Err(e) = audio_checker::parse_audio_file(file) {
                         file_entry.error_string = e.to_string();
                     }
                     Some(file_entry)
-                });
-
-                if let Ok(audio_result) = result {
-                    audio_result
-                } else {
+                })
+                .unwrap_or_else(|_| {
                     let message = create_crash_message("Symphonia", &file_entry_clone.path.to_string_lossy(), "https://github.com/pdeljanov/Symphonia");
                     println!("{message}");
                     file_entry_clone.error_string = message;
                     Some(file_entry_clone)
-                }
+                })
             }
             Err(_inspected) => None,
         }
     }
-    fn check_broken_pdf(&self, mut file_entry: FileEntry) -> Option<FileEntry> {
+    fn check_broken_pdf(&self, mut file_entry: BrokenEntry) -> Option<BrokenEntry> {
         let parser_options = ParseOptions::tolerant(); // Only show as broken files with really big bugs
 
         let mut file_entry_clone = file_entry.clone();
-        let result = panic::catch_unwind(|| {
+        panic::catch_unwind(|| {
             match FileOptions::cached().parse_options(parser_options).open(&file_entry.path) {
                 Ok(file) => {
                     for idx in 0..file.num_pages() {
@@ -308,33 +254,32 @@ impl BrokenFiles {
                 }
             }
             Some(file_entry)
-        });
-        if let Ok(pdf_result) = result {
-            pdf_result
-        } else {
+        })
+        .unwrap_or_else(|_| {
             let message = create_crash_message("PDF-rs", &file_entry_clone.path.to_string_lossy(), "https://github.com/pdf-rs/pdf");
             println!("{message}");
             file_entry_clone.error_string = message;
             Some(file_entry_clone)
-        }
+        })
     }
 
     #[fun_time(message = "load_cache", level = "debug")]
-    fn load_cache(&mut self) -> (BTreeMap<String, FileEntry>, BTreeMap<String, FileEntry>, BTreeMap<String, FileEntry>) {
+    fn load_cache(&mut self) -> (BTreeMap<String, BrokenEntry>, BTreeMap<String, BrokenEntry>, BTreeMap<String, BrokenEntry>) {
         let loaded_hash_map;
 
-        let mut records_already_cached: BTreeMap<String, FileEntry> = Default::default();
-        let mut non_cached_files_to_check: BTreeMap<String, FileEntry> = Default::default();
+        let mut records_already_cached: BTreeMap<String, BrokenEntry> = Default::default();
+        let mut non_cached_files_to_check: BTreeMap<String, BrokenEntry> = Default::default();
         let files_to_check = mem::take(&mut self.files_to_check);
 
         if self.common_data.use_cache {
-            let (messages, loaded_items) = load_cache_from_file_generalized_by_path::<FileEntry>(&get_broken_files_cache_file(), self.get_delete_outdated_cache(), &files_to_check);
+            let (messages, loaded_items) =
+                load_cache_from_file_generalized_by_path::<BrokenEntry>(&get_broken_files_cache_file(), self.get_delete_outdated_cache(), &files_to_check);
             self.get_text_messages_mut().extend_with_another_messages(messages);
             loaded_hash_map = loaded_items.unwrap_or_default();
 
             for (name, file_entry) in files_to_check {
                 if let Some(cached_file_entry) = loaded_hash_map.get(&name) {
-                    records_already_cached.insert(name.clone(), cached_file_entry.clone());
+                    records_already_cached.insert(name, cached_file_entry.clone());
                 } else {
                     non_cached_files_to_check.insert(name, file_entry);
                 }
@@ -347,14 +292,14 @@ impl BrokenFiles {
     }
 
     #[fun_time(message = "look_for_broken_files", level = "debug")]
-    fn look_for_broken_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
+    fn look_for_broken_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
         let (loaded_hash_map, records_already_cached, non_cached_files_to_check) = self.load_cache();
 
         let (progress_thread_handle, progress_thread_run, atomic_counter, _check_was_stopped) =
             prepare_thread_handler_common(progress_sender, 1, 1, non_cached_files_to_check.len(), CheckingMethod::None, self.common_data.tool_type);
 
         debug!("look_for_broken_files - started finding for broken files");
-        let mut vec_file_entry: Vec<FileEntry> = non_cached_files_to_check
+        let mut vec_file_entry: Vec<BrokenEntry> = non_cached_files_to_check
             .into_par_iter()
             .map(|(_, file_entry)| {
                 atomic_counter.fetch_add(1, Ordering::Relaxed);
@@ -374,7 +319,7 @@ impl BrokenFiles {
             .while_some()
             .filter(Option::is_some)
             .map(Option::unwrap)
-            .collect::<Vec<FileEntry>>();
+            .collect::<Vec<BrokenEntry>>();
         debug!("look_for_broken_files - ended finding for broken files");
 
         send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
@@ -397,10 +342,10 @@ impl BrokenFiles {
         true
     }
     #[fun_time(message = "save_to_cache", level = "debug")]
-    fn save_to_cache(&mut self, vec_file_entry: &[FileEntry], loaded_hash_map: BTreeMap<String, FileEntry>) {
+    fn save_to_cache(&mut self, vec_file_entry: &[BrokenEntry], loaded_hash_map: BTreeMap<String, BrokenEntry>) {
         if self.common_data.use_cache {
             // Must save all results to file, old loaded from file with all currently counted results
-            let mut all_results: BTreeMap<String, FileEntry> = Default::default();
+            let mut all_results: BTreeMap<String, BrokenEntry> = Default::default();
 
             for file_entry in vec_file_entry.iter().cloned() {
                 all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry);
@@ -420,7 +365,7 @@ impl BrokenFiles {
             DeleteMethod::Delete => {
                 for file_entry in &self.broken_files {
                     if fs::remove_file(&file_entry.path).is_err() {
-                        self.common_data.text_messages.warnings.push(file_entry.path.display().to_string());
+                        self.common_data.text_messages.warnings.push(file_entry.path.to_string_lossy().to_string());
                     }
                 }
             }
@@ -435,7 +380,7 @@ impl BrokenFiles {
 }
 
 impl BrokenFiles {
-    pub const fn get_broken_files(&self) -> &Vec<FileEntry> {
+    pub const fn get_broken_files(&self) -> &Vec<BrokenEntry> {
         &self.broken_files
     }
 
@@ -467,13 +412,15 @@ impl PrintResults for BrokenFiles {
         writeln!(
             writer,
             "Results of searching {:?} with excluded directories {:?} and excluded items {:?}",
-            self.common_data.directories.included_directories, self.common_data.directories.excluded_directories, self.common_data.excluded_items.items
+            self.common_data.directories.included_directories,
+            self.common_data.directories.excluded_directories,
+            self.common_data.excluded_items.get_excluded_items()
         )?;
 
         if !self.broken_files.is_empty() {
             writeln!(writer, "Found {} broken files.", self.information.number_of_broken_files)?;
             for file_entry in &self.broken_files {
-                writeln!(writer, "{} - {}", file_entry.path.display(), file_entry.error_string)?;
+                writeln!(writer, "{:?} - {}", file_entry.path, file_entry.error_string)?;
             }
         } else {
             write!(writer, "Not found any broken files.")?;
@@ -487,25 +434,37 @@ impl PrintResults for BrokenFiles {
     }
 }
 
-fn check_extension_availability(file_name_lowercase: &str) -> TypeOfFile {
-    if IMAGE_RS_BROKEN_FILES_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+fn check_extension_availability(
+    full_name: &Path,
+    images_extensions: &HashSet<&&'static str>,
+    zip_extensions: &HashSet<&&'static str>,
+    audio_extensions: &HashSet<&&'static str>,
+    pdf_extensions: &HashSet<&&'static str>,
+) -> TypeOfFile {
+    let Some(extension) = full_name.extension() else {
+        debug_assert!(false, "Missing extension");
+        return TypeOfFile::Unknown;
+    };
+
+    let Some(extension_str) = extension.to_str() else {
+        debug_assert!(false, "Extension not really fully str");
+        return TypeOfFile::Unknown;
+    };
+    let extension_lowercase = extension_str.to_ascii_lowercase();
+
+    if images_extensions.contains(&extension_lowercase.as_str()) {
         TypeOfFile::Image
-    } else if ZIP_FILES_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+    } else if zip_extensions.contains(&extension_lowercase.as_str()) {
         TypeOfFile::ArchiveZip
-    } else if AUDIO_FILES_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+    } else if audio_extensions.contains(&extension_lowercase.as_str()) {
         TypeOfFile::Audio
-    } else if PDF_FILES_EXTENSIONS.iter().any(|e| file_name_lowercase.ends_with(e)) {
+    } else if pdf_extensions.contains(&extension_lowercase.as_str()) {
         TypeOfFile::PDF
     } else {
+        eprintln!("File with unknown extension: {full_name:?} - {extension_lowercase}");
+        debug_assert!(false, "File with unknown extension");
         TypeOfFile::Unknown
     }
-}
-
-fn check_if_file_extension_is_allowed(type_of_file: &TypeOfFile, checked_types: &CheckedTypes) -> bool {
-    ((*type_of_file == TypeOfFile::Image) && ((*checked_types & CheckedTypes::IMAGE) == CheckedTypes::IMAGE))
-        || ((*type_of_file == TypeOfFile::PDF) && ((*checked_types & CheckedTypes::PDF) == CheckedTypes::PDF))
-        || ((*type_of_file == TypeOfFile::ArchiveZip) && ((*checked_types & CheckedTypes::ARCHIVE) == CheckedTypes::ARCHIVE))
-        || ((*type_of_file == TypeOfFile::Audio) && ((*checked_types & CheckedTypes::AUDIO) == CheckedTypes::AUDIO))
 }
 
 fn unpack_pdf_error(e: PdfError) -> PdfError {
@@ -523,7 +482,7 @@ fn unpack_pdf_error(e: PdfError) -> PdfError {
     }
 }
 
-fn validate_pdf_error(file_entry: &mut FileEntry, e: PdfError) -> PdfError {
+fn validate_pdf_error(file_entry: &mut BrokenEntry, e: PdfError) -> PdfError {
     let mut error_string = e.to_string();
     // Workaround for strange error message https://github.com/qarmin/czkawka/issues/898
     if error_string.starts_with("Try at") {
