@@ -2,10 +2,9 @@ use std::path::{Path, PathBuf};
 #[cfg(target_family = "unix")]
 use std::{fs, os::unix::fs::MetadataExt};
 
-use crate::common::Common;
+use crate::common::normalize_windows_path;
 use crate::common_messages::Messages;
 use crate::flc;
-use crate::localizer_core::generate_translation_hashmap;
 
 #[derive(Debug, Clone, Default)]
 pub struct Directories {
@@ -22,8 +21,21 @@ impl Directories {
         Default::default()
     }
 
-    pub fn set_reference_directory(&mut self, reference_directory: Vec<PathBuf>) {
-        self.reference_directories = reference_directory;
+    pub fn set_reference_directory(&mut self, reference_directory: &[PathBuf]) -> Messages {
+        let mut messages: Messages = Messages::new();
+
+        self.reference_directories = reference_directory
+            .iter()
+            .filter_map(|directory| {
+                let (dir, msg) = Self::canonicalize_and_clear_path(directory, false);
+
+                messages.extend_with_another_messages(msg);
+
+                dir
+            })
+            .collect::<Vec<PathBuf>>();
+
+        messages
     }
 
     pub fn set_included_directory(&mut self, included_directory: Vec<PathBuf>) -> Messages {
@@ -37,44 +49,14 @@ impl Directories {
         let directories: Vec<PathBuf> = included_directory;
 
         let mut checked_directories: Vec<PathBuf> = Vec::new();
-        for mut directory in directories {
-            if directory.to_string_lossy().contains('*') {
-                messages.warnings.push(flc!(
-                    "core_directory_wildcard_no_supported",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
+        for directory in directories {
+            let (dir, msg) = Self::canonicalize_and_clear_path(&directory, false);
 
-            if !directory.exists() {
-                messages.warnings.push(flc!(
-                    "core_directory_must_exists",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
-            if !directory.is_dir() {
-                messages.warnings.push(flc!(
-                    "core_directory_must_be_directory",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
+            messages.extend_with_another_messages(msg);
 
-            // If not checking windows strange paths, try to canonicalize them
-            if !directory.starts_with("\\") {
-                let Ok(dir2) = directory.canonicalize() else {
-                    messages.warnings.push(flc!(
-                        "core_directory_must_exists",
-                        generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                    ));
-                    continue;
-                };
-
-                directory = dir2;
+            if let Some(dir) = dir {
+                checked_directories.push(dir);
             }
-
-            checked_directories.push(directory);
         }
 
         if checked_directories.is_empty() {
@@ -103,46 +85,55 @@ impl Directories {
                 messages.errors.push(flc!("core_excluded_directory_pointless_slash"));
                 break;
             }
-            if directory_as_string.contains('*') {
-                messages.warnings.push(flc!(
-                    "core_directory_wildcard_no_supported",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
-            #[cfg(not(target_family = "windows"))]
-            if directory.is_relative() {
-                messages.warnings.push(flc!(
-                    "core_directory_relative_path",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
-            #[cfg(target_family = "windows")]
-            if directory.is_relative() && !directory.starts_with("\\") {
-                messages.warnings.push(flc!(
-                    "core_directory_relative_path",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
 
-            if !directory.exists() {
-                // No error when excluded directories are missing
-                continue;
+            let (dir, msg) = Self::canonicalize_and_clear_path(&directory, true);
+
+            messages.extend_with_another_messages(msg);
+
+            if let Some(dir) = dir {
+                checked_directories.push(dir);
             }
-            if !directory.is_dir() {
-                messages.warnings.push(flc!(
-                    "core_directory_must_be_directory",
-                    generate_translation_hashmap(vec![("path", directory.display().to_string())])
-                ));
-                continue;
-            }
-            checked_directories.push(directory);
         }
         self.excluded_directories = checked_directories;
 
         messages
+    }
+
+    fn canonicalize_and_clear_path(directory: &Path, is_excluded: bool) -> (Option<PathBuf>, Messages) {
+        let mut messages = Messages::new();
+        let mut directory = directory.to_path_buf();
+        if !directory.exists() {
+            if !is_excluded {
+                messages.warnings.push(flc!("core_directory_must_exists", path = directory.to_string_lossy().to_string()));
+            }
+            return (None, messages);
+        }
+
+        if !directory.is_dir() {
+            messages
+                .warnings
+                .push(flc!("core_directory_must_be_directory", path = directory.to_string_lossy().to_string()));
+            return (None, messages);
+        }
+
+        // Try to canonicalize them
+        if cfg!(windows) {
+            // Only canonicalize if it's not a network path
+            // This can be check by checking if path starts with \\?\UNC\
+            if let Ok(dir_can) = directory.canonicalize() {
+                let dir_can_str = dir_can.to_string_lossy().to_string();
+                if let Some(dir_can_str) = dir_can_str.strip_prefix(r"\\?\") {
+                    if dir_can_str.chars().nth(1) == Some(':') {
+                        directory = PathBuf::from(dir_can_str);
+                    }
+                }
+            }
+        } else {
+            if let Ok(dir) = directory.canonicalize() {
+                directory = dir;
+            }
+        }
+        (Some(directory), messages)
     }
 
     #[cfg(target_family = "unix")]
@@ -157,9 +148,9 @@ impl Directories {
         let mut optimized_excluded: Vec<PathBuf> = Vec::new();
 
         if cfg!(target_family = "windows") {
-            self.included_directories = self.included_directories.iter().map(Common::normalize_windows_path).collect();
-            self.excluded_directories = self.excluded_directories.iter().map(Common::normalize_windows_path).collect();
-            self.reference_directories = self.reference_directories.iter().map(Common::normalize_windows_path).collect();
+            self.included_directories = self.included_directories.iter().map(normalize_windows_path).collect();
+            self.excluded_directories = self.excluded_directories.iter().map(normalize_windows_path).collect();
+            self.reference_directories = self.reference_directories.iter().map(normalize_windows_path).collect();
         }
 
         // Remove duplicated entries like: "/", "/"
@@ -294,10 +285,7 @@ impl Directories {
             for d in &self.included_directories {
                 match fs::metadata(d) {
                     Ok(m) => self.included_dev_ids.push(m.dev()),
-                    Err(_) => messages.errors.push(flc!(
-                        "core_directory_unable_to_get_device_id",
-                        generate_translation_hashmap(vec![("path", d.display().to_string())])
-                    )),
+                    Err(_) => messages.errors.push(flc!("core_directory_unable_to_get_device_id", path = d.to_string_lossy().to_string())),
                 }
             }
         }
@@ -309,10 +297,9 @@ impl Directories {
         self.reference_directories.iter().any(|e| path.starts_with(e))
     }
 
-    pub fn is_excluded(&self, path: impl AsRef<Path>) -> bool {
-        let path = path.as_ref();
+    pub fn is_excluded(&self, path: &Path) -> bool {
         #[cfg(target_family = "windows")]
-        let path = Common::normalize_windows_path(path);
+        let path = normalize_windows_path(path);
         // We're assuming that `excluded_directories` are already normalized
         self.excluded_directories.iter().any(|p| p.as_path() == path)
     }
@@ -327,10 +314,7 @@ impl Directories {
         let path = path.as_ref();
         match fs::metadata(path) {
             Ok(m) => Ok(!self.included_dev_ids.iter().any(|&id| id == m.dev())),
-            Err(_) => Err(flc!(
-                "core_directory_unable_to_get_device_id",
-                generate_translation_hashmap(vec![("path", path.display().to_string())])
-            )),
+            Err(_) => Err(flc!("core_directory_unable_to_get_device_id", path = path.to_string_lossy().to_string())),
         }
     }
 }
