@@ -1,6 +1,14 @@
 use std::path::Path;
 
-use crate::common::Common;
+#[cfg(not(target_family = "unix"))]
+use crate::common::normalize_windows_path;
+use crate::common::regex_check;
+use crate::common_messages::Messages;
+
+#[cfg(target_family = "unix")]
+pub const DEFAULT_EXCLUDED_DIRECTORIES: &[&str] = &["/proc", "/dev", "/sys", "/run", "/snap"];
+#[cfg(not(target_family = "unix"))]
+pub const DEFAULT_EXCLUDED_DIRECTORIES: &[&str] = &["C:\\Windows"];
 
 #[cfg(target_family = "unix")]
 pub const DEFAULT_EXCLUDED_ITEMS: &str = "*/.git/*,*/node_modules/*,*/lost+found/*,*/Trash/*,*/.Trash-*/*,*/snap/*,/home/*/.cache/*";
@@ -9,7 +17,15 @@ pub const DEFAULT_EXCLUDED_ITEMS: &str = "*\\.git\\*,*\\node_modules\\*,*\\lost+
 
 #[derive(Debug, Clone, Default)]
 pub struct ExcludedItems {
-    pub items: Vec<String>,
+    expressions: Vec<String>,
+    connected_expressions: Vec<SingleExcludedItem>,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct SingleExcludedItem {
+    pub expression: String,
+    pub expression_splits: Vec<String>,
+    pub unique_extensions_splits: Vec<String>,
 }
 
 impl ExcludedItems {
@@ -17,12 +33,16 @@ impl ExcludedItems {
         Default::default()
     }
 
-    pub fn set_excluded_items(&mut self, excluded_items: Vec<String>) -> (Vec<String>, Vec<String>, Vec<String>) {
-        let messages: Vec<String> = Vec::new();
+    pub fn new_from(excluded_items: Vec<String>) -> Self {
+        let mut s = Self::new();
+        s.set_excluded_items(excluded_items);
+        s
+    }
+
+    pub fn set_excluded_items(&mut self, excluded_items: Vec<String>) -> Messages {
         let mut warnings: Vec<String> = Vec::new();
-        let errors: Vec<String> = Vec::new();
         if excluded_items.is_empty() {
-            return (messages, warnings, errors);
+            return Messages::new();
         }
 
         let expressions: Vec<String> = excluded_items;
@@ -49,19 +69,50 @@ impl ExcludedItems {
 
             checked_expressions.push(expression);
         }
-        self.items = checked_expressions;
-        (messages, warnings, errors)
+
+        for checked_expression in &checked_expressions {
+            let item = new_excluded_item(checked_expression);
+            self.expressions.push(item.expression.clone());
+            self.connected_expressions.push(item);
+        }
+        Messages {
+            messages: vec![],
+            warnings,
+            errors: vec![],
+        }
     }
 
-    pub fn is_excluded(&self, path: impl AsRef<Path>) -> bool {
+    pub fn get_excluded_items(&self) -> &Vec<String> {
+        &self.expressions
+    }
+    pub fn is_excluded(&self, path: &Path) -> bool {
+        if self.connected_expressions.is_empty() {
+            return false;
+        }
         #[cfg(target_family = "windows")]
-        let path = Common::normalize_windows_path(path);
+        let path = normalize_windows_path(path);
 
-        for expression in &self.items {
-            if Common::regex_check(expression, &path) {
+        let path_str = path.to_string_lossy();
+
+        for expression in &self.connected_expressions {
+            if regex_check(expression, &path_str) {
                 return true;
             }
         }
         false
+    }
+}
+
+pub fn new_excluded_item(expression: &str) -> SingleExcludedItem {
+    let expression = expression.trim().to_string();
+    let expression_splits: Vec<String> = expression.split('*').filter_map(|e| if e.is_empty() { None } else { Some(e.to_string()) }).collect();
+    let mut unique_extensions_splits = expression_splits.clone();
+    unique_extensions_splits.sort();
+    unique_extensions_splits.dedup();
+    unique_extensions_splits.sort_by_key(|b| std::cmp::Reverse(b.len()));
+    SingleExcludedItem {
+        expression,
+        expression_splits,
+        unique_extensions_splits,
     }
 }
