@@ -1,19 +1,17 @@
 use std::fs;
-use std::fs::{DirEntry, Metadata};
+use std::fs::DirEntry;
 use std::io::prelude::*;
-
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 
-use crossbeam_channel::Receiver;
+use crossbeam_channel::{Receiver, Sender};
 use fun_time::fun_time;
-use futures::channel::mpsc::UnboundedSender;
 use rayon::prelude::*;
 use serde::Serialize;
 
 use crate::common::{check_folder_children, check_if_stop_received, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads};
-use crate::common_dir_traversal::{common_get_entry_data_metadata, common_read_dir, get_lowercase_name, get_modified_time, CheckingMethod, ProgressData, ToolType};
+use crate::common_dir_traversal::{common_read_dir, get_modified_time, CheckingMethod, ProgressData, ToolType};
 use crate::common_tool::{CommonData, CommonToolData, DeleteMethod};
 use crate::common_traits::*;
 
@@ -33,10 +31,19 @@ const TEMP_EXTENSIONS: &[&str] = &[
     ".partial",
 ];
 
-#[derive(Clone, Serialize)]
-pub struct FileEntry {
+#[derive(Clone, Serialize, Debug)]
+pub struct TemporaryFileEntry {
     pub path: PathBuf,
     pub modified_date: u64,
+}
+
+impl TemporaryFileEntry {
+    pub fn get_path(&self) -> &PathBuf {
+        &self.path
+    }
+    pub fn get_modified_date(&self) -> u64 {
+        self.modified_date
+    }
 }
 
 #[derive(Default)]
@@ -47,7 +54,7 @@ pub struct Info {
 pub struct Temporary {
     common_data: CommonToolData,
     information: Info,
-    temporary_files: Vec<FileEntry>,
+    temporary_files: Vec<TemporaryFileEntry>,
 }
 
 impl Temporary {
@@ -60,8 +67,8 @@ impl Temporary {
     }
 
     #[fun_time(message = "find_temporary_files", level = "info")]
-    pub fn find_temporary_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) {
-        self.optimize_dirs_before_start();
+    pub fn find_temporary_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) {
+        self.prepare_items();
         if !self.check_files(stop_receiver, progress_sender) {
             self.common_data.stopped_search = true;
             return;
@@ -71,13 +78,8 @@ impl Temporary {
     }
 
     #[fun_time(message = "check_files", level = "debug")]
-    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&UnboundedSender<ProgressData>>) -> bool {
-        let mut folders_to_check: Vec<PathBuf> = Vec::with_capacity(1024 * 2); // This should be small enough too not see to big difference and big enough to store most of paths without needing to resize vector
-
-        // Add root folders for finding
-        for id in &self.common_data.directories.included_directories {
-            folders_to_check.push(id.clone());
-        }
+    fn check_files(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
+        let mut folders_to_check: Vec<PathBuf> = self.common_data.directories.included_directories.clone();
 
         let (progress_thread_handle, progress_thread_run, atomic_counter, _check_was_stopped) =
             prepare_thread_handler_common(progress_sender, 0, 0, 0, CheckingMethod::None, self.common_data.tool_type);
@@ -89,34 +91,36 @@ impl Temporary {
             }
 
             let segments: Vec<_> = folders_to_check
-                .par_iter()
+                .into_par_iter()
                 .map(|current_folder| {
                     let mut dir_result = vec![];
                     let mut warnings = vec![];
                     let mut fe_result = vec![];
 
-                    let Some(read_dir) = common_read_dir(current_folder, &mut warnings) else {
+                    let Some(read_dir) = common_read_dir(&current_folder, &mut warnings) else {
                         return (dir_result, warnings, fe_result);
                     };
 
                     // Check every sub folder/file/link etc.
                     for entry in read_dir {
-                        let Some((entry_data, metadata)) = common_get_entry_data_metadata(&entry, &mut warnings, current_folder) else {
+                        let Ok(entry_data) = entry else {
+                            continue;
+                        };
+                        let Ok(file_type) = entry_data.file_type() else {
                             continue;
                         };
 
-                        if metadata.is_dir() {
+                        if file_type.is_dir() {
                             check_folder_children(
                                 &mut dir_result,
                                 &mut warnings,
-                                current_folder,
-                                entry_data,
+                                &entry_data,
                                 self.common_data.recursive_search,
                                 &self.common_data.directories,
                                 &self.common_data.excluded_items,
                             );
-                        } else if metadata.is_file() {
-                            if let Some(file_entry) = self.get_file_entry(&metadata, &atomic_counter, entry_data, &mut warnings, current_folder) {
+                        } else if file_type.is_file() {
+                            if let Some(file_entry) = self.get_file_entry(&atomic_counter, &entry_data, &mut warnings) {
                                 fe_result.push(file_entry);
                             }
                         }
@@ -125,8 +129,8 @@ impl Temporary {
                 })
                 .collect();
 
-            // Advance the frontier
-            folders_to_check.clear();
+            let required_size = segments.iter().map(|(segment, _, _)| segment.len()).sum::<usize>();
+            folders_to_check = Vec::with_capacity(required_size);
 
             // Process collected data
             for (segment, warnings, fe_result) in segments {
@@ -143,32 +147,29 @@ impl Temporary {
 
         true
     }
-    pub fn get_file_entry(
-        &self,
-        metadata: &Metadata,
-        atomic_counter: &Arc<AtomicUsize>,
-        entry_data: &DirEntry,
-        warnings: &mut Vec<String>,
-        current_folder: &Path,
-    ) -> Option<FileEntry> {
+    pub fn get_file_entry(&self, atomic_counter: &Arc<AtomicUsize>, entry_data: &DirEntry, warnings: &mut Vec<String>) -> Option<TemporaryFileEntry> {
         atomic_counter.fetch_add(1, Ordering::Relaxed);
 
-        let Some(file_name_lowercase) = get_lowercase_name(entry_data, warnings) else {
-            return None;
-        };
-
-        if !TEMP_EXTENSIONS.iter().any(|f| file_name_lowercase.ends_with(f)) {
-            return None;
-        }
-        let current_file_name = current_folder.join(entry_data.file_name());
+        let current_file_name = entry_data.path();
         if self.common_data.excluded_items.is_excluded(&current_file_name) {
             return None;
         }
 
+        let file_name = entry_data.file_name();
+        let file_name_ascii_lowercase = file_name.to_ascii_lowercase();
+        let file_name_lowercase = file_name_ascii_lowercase.to_string_lossy();
+        if !TEMP_EXTENSIONS.iter().any(|f| file_name_lowercase.ends_with(f)) {
+            return None;
+        }
+
+        let Ok(metadata) = entry_data.metadata() else {
+            return None;
+        };
+
         // Creating new file entry
-        Some(FileEntry {
-            path: current_file_name.clone(),
-            modified_date: get_modified_time(metadata, warnings, &current_file_name, false),
+        Some(TemporaryFileEntry {
+            modified_date: get_modified_time(&metadata, warnings, &current_file_name, false),
+            path: current_file_name,
         })
     }
 
@@ -179,7 +180,7 @@ impl Temporary {
                 let mut warnings = Vec::new();
                 for file_entry in &self.temporary_files {
                     if fs::remove_file(file_entry.path.clone()).is_err() {
-                        warnings.push(file_entry.path.display().to_string());
+                        warnings.push(file_entry.path.to_string_lossy().to_string());
                     }
                 }
                 self.common_data.text_messages.warnings.extend(warnings);
@@ -197,12 +198,14 @@ impl PrintResults for Temporary {
         writeln!(
             writer,
             "Results of searching {:?} with excluded directories {:?} and excluded items {:?}",
-            self.common_data.directories.included_directories, self.common_data.directories.excluded_directories, self.common_data.excluded_items.items
+            self.common_data.directories.included_directories,
+            self.common_data.directories.excluded_directories,
+            self.common_data.excluded_items.get_excluded_items()
         )?;
         writeln!(writer, "Found {} temporary files.\n", self.information.number_of_temporary_files)?;
 
         for file_entry in &self.temporary_files {
-            writeln!(writer, "{}", file_entry.path.display())?;
+            writeln!(writer, "{:?}", file_entry.path)?;
         }
 
         Ok(())
@@ -240,7 +243,7 @@ impl CommonData for Temporary {
 }
 
 impl Temporary {
-    pub const fn get_temporary_files(&self) -> &Vec<FileEntry> {
+    pub const fn get_temporary_files(&self) -> &Vec<TemporaryFileEntry> {
         &self.temporary_files
     }
 
