@@ -33,18 +33,22 @@ impl GuiAbout {
             "Rafał Mikrut",
             "Alexis Lefebvre",
             "Thomas Andreas Jung",
+            "Alexis Lefebvre",
             "Peter Blackson",
+            "qarmin",
             "TheEvilSkeleton",
             "Ben Bodenmiller",
             "ChihWei Wang",
             "Dan Dascalescu",
             "Dominik Piątkowski",
             "Igor",
+            "Jocelyn Le Sage",
             "Kerfuffle",
             "Shriraj Hegde",
             "krzysdz",
             "0x4A6F",
             "0xflotus",
+            "Aarni Koskela",
             "Adam Boguszewski",
             "Alex",
             "Caduser2020",
@@ -53,6 +57,7 @@ impl GuiAbout {
             "Douman",
             "Elazar Fine",
             "Farmadupe",
+            "Fr_Dae",
             "Gitoffthelawn",
             "Ivan Habernal",
             "Jan Jurec",
@@ -60,9 +65,11 @@ impl GuiAbout {
             "Jona",
             "Jonathan Hult",
             "Kian-Meng Ang",
+            "Liru Wilkowski",
             "Meir Klemfner",
             "Mek101",
             "Michael Grigoryan",
+            "Nick Gallimore",
             "Nikita Karamov",
             "OMEGA_RAZER",
             "Rodrigo Torres",
@@ -70,17 +77,23 @@ impl GuiAbout {
             "Spirit",
             "Stefan Seering",
             "Syfaro",
+            "Tom Paine",
+            "Tom Praschan",
             "Yuri Slobodyanyuk",
             "alexdraconian",
             "bakeromso",
             "bellrise",
+            "codingnewcode",
             "cyqsimon",
             "endolith",
             "jann",
             "kamilek96",
             "kuskov",
+            "rugk",
+            "santiago fn",
             "tecome",
             "tenninjas",
+            "undefined-landmark",
         ]);
 
         let custom_box = get_all_boxes_from_widget(&about_dialog)[2].clone(); // TODO may not be stable enough between GTK versions
@@ -108,7 +121,7 @@ impl GuiAbout {
     }
 
     pub fn update_language(&self) {
-        let mut comment_text: String = "2020 - 2023  Rafał Mikrut(qarmin)\n\n".to_string();
+        let mut comment_text: String = "2020 - 2024  Rafał Mikrut(qarmin)\n\n".to_string();
         comment_text += &flg!("about_window_motto");
         self.about_dialog.set_comments(Some(&comment_text));
 
